refactoring
has bugs, needs work to function more concisely and readably. very oop solutions on the way.
specifically, creating an "answer" class and data type to reutrn after solve sum so that it can return skips and max total
also want to create tester that addresses directly to solveSum method and uses outputAnswers to output to console
to do:
- answer class, getters/setters
- tester class
- clean comments/printers
- make combo more readable
   - maybe extra method to "get combo"